malloc protected

+++ libft/ft_strncat.c
@@ -6,14 +6,13 @@
 /*   By: ibouabda <ibouabda@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/07/10 21:36:13 by ibouabda          #+#    #+#             */
-/*   Updated: 2018/11/27 14:45:20 by ibouabda         ###   ########.fr       */
+/*   Updated: 2019/08/29 13:39:07 by ibouabda         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <stdio.h>
-#include <string.h>
+#include "libft.h"
 
-char	*ft_strncat(char *dest, const char *src, int nb)
+char	*ft_strncat(char *dest, const char *src, size_t nb)
 {
 	int i;
 	int k;
@@ -24,7 +23,7 @@ char	*ft_strncat(char *dest, const char *src, int nb)
 	{
 		i++;
 	}
-	while (k < nb && src[k])
+	while (k < (int)nb && src[k])
 	{
 		dest[i + k] = src[k];
 		k++;
